Day 15 optimisations
I took a little bit of inspiration from others' answers to do this, but
basically it was removing as many comparisons/conditionals as possible.

- Change the nested dicts to a single-level dict, calculate the previous
instance as i-1 (but then I got rid of that because for calculating the
differences I can just use 'i' and the off-by-one doesn't matter as it's
on both sides. I had to increment the index on the starting range, but
that's a small price given we only have a few of those). I think this
gives the biggest runtime benefit.
- Replace the conditionals with an equation that satisfies either
conditions (yes, this sort of buries a conditional in the dict default
handler but lets just roll with it)
- Remove all the temporary named variables

This is all a bit golfy and ends up with some code that's harder to
follow back to the original game rules, but it does run in half the time
so. 🤷‍♂️